Move al redux related files to /src/redux
