+++ src/bootstrap.js
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 import { Provider, useDispatch, useSelector, useStore } from 'react-redux';
 import { IntlProvider, ReactIntlErrorCode } from 'react-intl';
-import { spinUpStore } from './js/redux-config';
+import { spinUpStore } from './redux/redux-config';
 import RootApp from './js/App/RootApp';
-import { loadModulesSchema } from './js/redux/actions';
+import { loadModulesSchema } from './redux/actions';
 import Cookies from 'js-cookie';
 import { ACTIVE_REMOTE_REQUEST, CROSS_ACCESS_ACCOUNT_NUMBER } from './js/consts';
 import auth, { crossAccountBouncer } from './js/auth';
