+++ src/js/App/GlobalFilter/TagsModal.tsx
@@ -2,12 +2,12 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 import { shallowEqual, useDispatch, useSelector } from 'react-redux';
 import { TagModal } from '@redhat-cloud-services/frontend-components/TagModal';
-import { fetchAllSIDs, fetchAllTags } from '../../redux/actions';
+import { fetchAllSIDs, fetchAllTags } from '../../../redux/actions';
 import debounce from 'lodash/debounce';
 import flatMap from 'lodash/flatMap';
 import { useIntl } from 'react-intl';
 import messages from '../../Messages';
-import { CommonSelectedTag, CommonTag, GlobalFilterTag, ReduxState, SID } from '../../redux/store';
+import { CommonSelectedTag, CommonTag, GlobalFilterTag, ReduxState, SID } from '../../../redux/store';
 import { FlagTagsFilter } from './constants';
 import { TagFilterOptions, TagPagination } from './tagsApi';
 import { Action } from 'redux';
