+++ src/js/App/Sidenav/LandingNav.js
@@ -4,7 +4,7 @@ import { isBeta, isFedRamp } from '../../../utils/common';
 import './LandingNav.scss';
 import { useDispatch, useSelector } from 'react-redux';
 import axios from 'axios';
-import { loadNavigationLandingPage } from '../../redux/actions';
+import { loadNavigationLandingPage } from '../../../redux/actions';
 import NavLoader from './Navigation/Loader';
 import ChromeNavItemFactory from './Navigation/ChromeNavItemFactory';
 import NavContext from './Navigation/navContext';
