+++ src/redux/chromeReducers.test.js
@@ -1,6 +1,5 @@
-import '../test';
 import * as actions from './actions';
-import * as reducers from './reducers';
+import * as reducers from './chromeReducers';
 
 const mockNav = [
   {
