+++ src/redux/globalFilterReducers.ts
